Refactored application entry point and created translation file
